"low changes"

+++ trunk/code/ElectronicElections/src/RSA/RSACipher.h
@@ -22,7 +22,7 @@ public:
 	int64 getRelativelyPrimeNumber(int64 number);
 	static int64 productInverse(int64 n, int64 d);
 	static int64 modularExponentiation(int64 a, int64 b, int64 q); // a^b mod q
-	void cipherMessage(char* message, int expKey, int64 n, char* cipheredMessage);
+	void cipherMessage(char* message, int64 expKey, int64 n, char* cipheredMessage);
 	virtual ~RSACipher();
 };
